requirements.txt
asgiref==3.9.1
crispy-bootstrap5==2025.6
Django==5.2.6
django-crispy-forms==2.4
dj-database-url==2.2.0
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
python-dotenv==1.0.0
reportlab==4.2.5
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
